Add pretty print to perftimer
